requirements.txt
python-telegram-bot>=21.10
google-genai>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
aiofiles>=23.0.0
lxml>=5.0.0
